solver: change builder.call to builder.context
