+++ solver/jobs.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/moby/buildkit/client"
-	"github.com/moby/buildkit/identity"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/util/flightcontrol"
 	"github.com/moby/buildkit/util/progress"
@@ -21,7 +20,7 @@ type ResolveOpFunc func(Vertex, Builder) (Op, error)
 
 type Builder interface {
 	Build(ctx context.Context, e Edge) (CachedResult, error)
-	Call(ctx context.Context, name string, fn func(ctx context.Context) error) error
+	Context(ctx context.Context) context.Context
 }
 
 // Solver provides a shared graph of all the vertexes currently being
@@ -166,9 +165,8 @@ func (sb *subBuilder) Build(ctx context.Context, e Edge) (CachedResult, error) {
 	return res, nil
 }
 
-func (sb *subBuilder) Call(ctx context.Context, name string, fn func(ctx context.Context) error) error {
-	ctx = progress.WithProgress(ctx, sb.mpw)
-	return inVertexContext(ctx, name, fn)
+func (sb *subBuilder) Context(ctx context.Context) context.Context {
+	return progress.WithProgress(ctx, sb.mpw)
 }
 
 type Job struct {
@@ -444,9 +442,8 @@ func (j *Job) Discard() error {
 	return nil
 }
 
-func (j *Job) Call(ctx context.Context, name string, fn func(ctx context.Context) error) error {
-	ctx = progress.WithProgress(ctx, j.pw)
-	return inVertexContext(ctx, name, fn)
+func (j *Job) Context(ctx context.Context) context.Context {
+	return progress.WithProgress(ctx, j.pw)
 }
 
 type cacheMapResp struct {
@@ -761,17 +758,4 @@ func notifyCompleted(ctx context.Context, v *client.Vertex, err error, cached bo
 		v.Error = err.Error()
 	}
 	pw.Write(v.Digest.String(), *v)
-}
-
-func inVertexContext(ctx context.Context, name string, f func(ctx context.Context) error) error {
-	v := client.Vertex{
-		Digest: digest.FromBytes([]byte(identity.NewID())),
-		Name:   name,
-	}
-	pw, _, ctx := progress.FromContext(ctx, progress.WithMetadata("vertex", v.Digest))
-	notifyStarted(ctx, &v, false)
-	defer pw.Close()
-	err := f(ctx)
-	notifyCompleted(ctx, &v, err, false)
-	return err
 }
